requirements.txt
pillow
pyautogui
requests